Move around files related to zip deploy
