+++ appservice/src/deploy/deployZip.ts
@@ -3,19 +3,20 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AppServicePlan, StringDictionary } from 'azure-arm-website/lib/models';
+import { AppServicePlan } from 'azure-arm-website/lib/models';
 import * as fs from 'fs';
 import * as fse from 'fs-extra';
-import * as path from 'path';
 import * as vscode from 'vscode';
-import { IActionContext } from 'vscode-azureextensionui';
 import { KuduClient } from 'vscode-azurekudu';
 import { ext } from '../extensionVariables';
-import * as FileUtilities from '../FileUtilities';
 import { localize } from '../localize';
 import { SiteClient } from '../SiteClient';
+import { getFileExtension } from '../utils/pathUtils';
+import { delayFirstWebAppDeploy } from './delayFirstWebAppDeploy';
 import { deployToStorageAccount } from './deployToStorageAccount';
+import { getZipFileToDeploy } from './getZipFileToDeploy';
 import { IDeployContext } from './IDeployContext';
+import { validateLinuxFunctionAppSettings } from './validateLinuxFunctionAppSettings';
 import { waitForDeploymentToComplete } from './waitForDeploymentToComplete';
 
 export async function deployZip(context: IDeployContext, client: SiteClient, fsPath: string, aspPromise: Promise<AppServicePlan | undefined>): Promise<void> {
@@ -25,7 +26,7 @@ export async function deployZip(context: IDeployContext, client: SiteClient, fsP
 
     let zipFilePath: string;
     let createdZip: boolean = false;
-    if (FileUtilities.getFileExtension(fsPath) === 'zip') {
+    if (getFileExtension(fsPath) === 'zip') {
         context.telemetry.properties.alreadyZipped = 'true';
         zipFilePath = fsPath;
     } else {
@@ -78,98 +79,5 @@ export async function deployZip(context: IDeployContext, client: SiteClient, fsP
         if (createdZip) {
             await fse.remove(zipFilePath);
         }
-    }
-}
-
-async function getZipFileToDeploy(fsPath: string, isFunctionApp: boolean): Promise<string> {
-    if ((await fse.lstat(fsPath)).isDirectory()) {
-        if (!fsPath.endsWith(path.sep)) {
-            fsPath += path.sep;
-        }
-
-        if (isFunctionApp) {
-            return FileUtilities.zipDirectoryGitignore(fsPath, '.funcignore');
-        } else {
-            const zipDeployConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration(ext.prefix, vscode.Uri.file(fsPath));
-            const globPattern: string | undefined = zipDeployConfig.get<string>('zipGlobPattern');
-            const ignorePattern: string | string[] | undefined = zipDeployConfig.get<string | string[]>('zipIgnorePattern');
-            return FileUtilities.zipDirectoryGlob(fsPath, globPattern, ignorePattern);
-        }
-    } else {
-        return FileUtilities.zipFile(fsPath);
-    }
-}
-
-async function delayFirstWebAppDeploy(client: SiteClient, asp: AppServicePlan | undefined): Promise<void> {
-    await new Promise<void>(async (resolve: () => void): Promise<void> => {
-        setTimeout(resolve, 10000);
-        try {
-            // this delay is only valid for the first deployment to a Linux web app on a basic asp, so resolve for anything else
-            if (client.isFunctionApp) {
-                resolve();
-            }
-            if (!asp || !asp.sku || !asp.sku.tier || asp.sku.tier.toLowerCase() !== 'basic') {
-                resolve();
-            }
-            if (!client.isLinux) {
-                resolve();
-            }
-
-            const kuduClient: KuduClient = await client.getKuduClient();
-            const deployments: number = (await kuduClient.deployment.getDeployResults()).length;
-            if (deployments > 1) {
-                resolve();
-            }
-        } catch (error) {
-            // ignore the error, an error here isn't a deployment failure
-            resolve();
-        }
-    });
-}
-
-async function validateLinuxFunctionAppSettings(context: IActionContext, client: SiteClient, doBuild: boolean, isConsumption: boolean): Promise<void> {
-    const appSettings: StringDictionary = await client.listApplicationSettings();
-    // tslint:disable-next-line:strict-boolean-expressions
-    appSettings.properties = appSettings.properties || {};
-
-    let hasChanged: boolean = false;
-
-    const keysToRemove: string[] = [];
-
-    if (doBuild) {
-        keysToRemove.push(
-            'WEBSITE_RUN_FROM_ZIP',
-            'WEBSITE_RUN_FROM_PACKAGE'
-        );
-    }
-
-    if (!isConsumption) {
-        const dedicatedBuildSettings: [string, string][] = [
-            ['ENABLE_ORYX_BUILD', 'true'],
-            ['SCM_DO_BUILD_DURING_DEPLOYMENT', '1'],
-            ['BUILD_FLAGS', 'UseExpressBuild'],
-            ['XDG_CACHE_HOME', '/tmp/.cache']
-        ];
-
-        for (const [key, value] of dedicatedBuildSettings) {
-            if (!doBuild) {
-                keysToRemove.push(key);
-            } else if (appSettings.properties[key] !== value) {
-                appSettings.properties[key] = value;
-                hasChanged = true;
-            }
-        }
-    }
-
-    for (const key of keysToRemove) {
-        if (appSettings.properties[key]) {
-            delete appSettings.properties[key];
-            hasChanged = true;
-        }
-    }
-
-    if (hasChanged) {
-        context.telemetry.properties.updatedAppSettings = 'true';
-        await client.updateApplicationSettings(appSettings);
     }
 }
